requirement.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
google-genai
numpy
scikit-learn